extract CityItem logic in seperate component
