+++ weather-app/src/Components/CityItem/CityItem.css
@@ -1,3 +1,9 @@
+.cityItemWrapper {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
 .CityItem {
     width: 90%;
     display: flex;
@@ -16,31 +22,3 @@
     width: fit-content;
 }
 
-.CityItemCardWrapper {
-    width: fit-content;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-}
-
-/* .CityItemCard {
-    background-color: #00008039;
-    border-radius: 20px;
-    padding: 0.5rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    min-width: 60px;
-}
-
-.CityItemCardMonth {
-    text-transform: uppercase;
-    font-size: 0.9rem;
-    height: 1.5rem
-}
-
-.CityItemCardIcon {
-    margin: 10px 0;
-} */
